2
2

+++ main.c
@@ -3,7 +3,7 @@
  
 int main()
 {
-   printf("int 存储大小 : %lu \n", sizeof(long));
+   printf("int 存储大小 : %lu \n", sizeof(int));
    
    return 0;
 }
